Extract renewal handler logic to new class
