+++ modules/ppcp-paypal-subscriptions/services.php
@@ -40,4 +40,7 @@ return array(
 			dirname( realpath( __FILE__ ), 3 ) . '/woocommerce-paypal-payments.php'
 		);
 	},
+	'paypal-subscriptions.renewal-handler'          => static function ( ContainerInterface $container ): RenewalHandler {
+		return new RenewalHandler( $container->get( 'woocommerce.logger.woocommerce' ) );
+	},
 );
